backend/requirements.txt
# Backend Requirements

# Core
fastapi
uvicorn
pydantic
pydantic-settings
httpx

# Infrastructure
ccxt>=4.3.0
sqlalchemy
psycopg2-binary
redis
chromadb
apscheduler

# Structlog for JSON logging
structlog
python-json-logger

# Data Analysis
pandas>=2.2.0
pandas-ta>=0.3.14b
numpy>=1.26.0

# AI & Machine Learning
openai
google-generativeai
anthropic
stable-baselines3>=2.3.0
gymnasium>=0.29.0
scikit-learn

# Auth & Security
python-jose[cryptography]
passlib[bcrypt]
pyotp
python-multipart
slowapi
